create confirmation email service and mock it for user.mutation test
